+++ crates/nixci/src/config.rs
@@ -22,22 +22,24 @@ use crate::cli::BuildConfig;
 ///     overrideInputs = { "mymod" = "."; };
 ///   };
 /// }
-// NB: we use BTreeMap instead of HashMap here so that we always iterate
-// configs in a determinitstic (i.e. asciibetical) order
 #[derive(Debug)]
 pub struct Config {
     /// The flake.nix configuration
-    pub subflakes: Subflakes,
+    pub subflakes: SubflakesConfig,
 
-    /// The URL to the flake containing this configuration
+    pub ref_: ConfigRef,
+}
+
+/// A reference into one or all [SubflakesConfig] of some [FlakeUrl]
+#[derive(Debug)]
+pub struct ConfigRef {
+    /// The flake itself
     pub flake_url: FlakeUrl,
 
-    /// Configuration name (`nixci.<name>`)
-    pub name: String,
+    /// The name of the nixci configuration (`omci.<name>`) selected
+    pub selected_name: String,
 
-    /// Selected sub-flake if any.
-    ///
-    /// Must be a key in `subflakes`.
+    /// The selected sub-flake if any.
     pub selected_subflake: Option<String>,
 }
 
@@ -51,9 +53,10 @@ impl Config {
     /// ```
     /// along with the config.
     pub async fn from_flake_url(cmd: &NixCmd, url: &FlakeUrl) -> Result<Config> {
-        let flake_attr = RootQualifiedAttr::new(&["om.ci", "nixci"]);
-        let (subflakes, flake_url, rest_attrs) =
-            flake_attr.eval_flake::<Subflakes>(cmd, url).await?;
+        let (mut subflakes, selected_name, rest_attrs) =
+            RootQualifiedAttr::new(&["om.ci", "nixci"])
+                .eval_flake::<SubflakesConfig>(cmd, url)
+                .await?;
         let selected_subflake = rest_attrs.first().cloned();
         if let Some(sub_flake_name) = selected_subflake.clone() {
             if !subflakes.0.contains_key(&sub_flake_name) {
@@ -63,26 +66,35 @@ impl Config {
                     url
                 )
             }
+            for (name, value) in subflakes.0.iter_mut() {
+                if name != &sub_flake_name {
+                    value.skip = true;
+                }
+            }
         }
-        let cfg = Config {
-            subflakes,
-            flake_url: flake_url.clone(),
-            name: flake_url.split_attr().1.get_name().to_string(),
+        let ref_ = ConfigRef {
+            flake_url: url.without_attr(),
+            selected_name,
             selected_subflake,
         };
+        let cfg = Config { subflakes, ref_ };
         Ok(cfg)
     }
 }
 
 #[derive(Debug, Deserialize)]
-pub struct Subflakes(pub BTreeMap<String, SubFlakish>);
+pub struct SubflakesConfig(
+    // NB: we use BTreeMap instead of HashMap here so that we always iterate
+    // configs in a determinitstic (i.e. asciibetical) order
+    pub BTreeMap<String, SubflakeConfig>,
+);
 
-impl Default for Subflakes {
+impl Default for SubflakesConfig {
     /// Default value contains a single entry for the root flake.
     fn default() -> Self {
         let mut subflakes = BTreeMap::new();
-        subflakes.insert("<root>".to_string(), SubFlakish::default());
-        Subflakes(subflakes)
+        subflakes.insert("<root>".to_string(), SubflakeConfig::default());
+        SubflakesConfig(subflakes)
     }
 }
 
@@ -91,7 +103,11 @@ impl Default for Subflakes {
 /// "Look-alike" because its inputs may be partial, thus requiring explicit
 /// --override-inputs when evaluating the flake.
 #[derive(Debug, Deserialize)]
-pub struct SubFlakish {
+pub struct SubflakeConfig {
+    /// Whether to skip building this subflake
+    #[serde(default)]
+    pub skip: bool,
+
     /// Subdirectory in which the flake lives
     pub dir: String,
 
@@ -105,10 +121,11 @@ pub struct SubFlakish {
     pub systems: Option<Vec<System>>,
 }
 
-impl Default for SubFlakish {
-    /// The default `SubFlakish` is the root flake.
+impl Default for SubflakeConfig {
+    /// The default `SubflakeConfig` is the root flake.
     fn default() -> Self {
-        SubFlakish {
+        SubflakeConfig {
+            skip: false,
             dir: ".".to_string(),
             override_inputs: BTreeMap::default(),
             systems: None,
@@ -116,7 +133,7 @@ impl Default for SubFlakish {
     }
 }
 
-impl SubFlakish {
+impl SubflakeConfig {
     pub fn can_build_on(&self, systems: &[System]) -> bool {
         match self.systems.as_ref() {
             Some(systems_whitelist) => systems_whitelist.iter().any(|s| systems.contains(s)),
@@ -167,8 +184,8 @@ mod tests {
         let cfg = Config::from_flake_url(&NixCmd::default(), url)
             .await
             .unwrap();
-        assert_eq!(cfg.name, "default");
-        assert_eq!(cfg.selected_subflake, Some("dev".to_string()));
+        assert_eq!(cfg.ref_.selected_name, "default");
+        assert_eq!(cfg.ref_.selected_subflake, Some("dev".to_string()));
         assert_eq!(cfg.subflakes.0.len(), 7);
     }
 }