+++ crates/nixci/src/github/matrix.rs
@@ -2,7 +2,7 @@
 use nix_rs::flake::system::System;
 use serde::{Deserialize, Serialize};
 
-use crate::config::Subflakes;
+use crate::config::SubflakesConfig;
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct GitHubMatrixRow {
@@ -16,7 +16,7 @@ pub struct GitHubMatrix {
 }
 
 impl GitHubMatrix {
-    pub fn from(systems: Vec<System>, subflakes: &Subflakes) -> Self {
+    pub fn from(systems: Vec<System>, subflakes: &SubflakesConfig) -> Self {
         let include: Vec<GitHubMatrixRow> = systems
             .iter()
             .flat_map(|system| {